requirements.txt
click==8.1.7
pydantic>=2.5
PyYAML>=6.0.1
openai>=1.35.0
rich>=13.7.0
requests>=2.32.0
pytest>=8.2.0